organized js folder, added chat animation
